Rename api directory to cmd/openfish
This makes it consistent with https://github.com/ausocean/cloud and the wider go community
