+++ .github/workflows/code-checks.yaml
@@ -14,7 +14,7 @@ jobs:
         go-version: '1.21'
 
     - name: Build
-      run: go build -v -o ./openfish-api ./api
+      run: go build -v ./cmd/openfish
 
     - name: Check formatting
       run: test -z $(gofmt -l .)
