+++ cmd/openfish/handlers/videostream.go
@@ -39,9 +39,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/ausocean/openfish/api/api"
-	"github.com/ausocean/openfish/api/entities"
-	"github.com/ausocean/openfish/api/services"
+	"github.com/ausocean/openfish/cmd/openfish/api"
+	"github.com/ausocean/openfish/cmd/openfish/entities"
+	"github.com/ausocean/openfish/cmd/openfish/services"
 
 	"github.com/gofiber/fiber/v2"
 )
