+++ cmd/openfish/middleware/authentication.go
@@ -37,9 +37,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/ausocean/openfish/api/api"
-	"github.com/ausocean/openfish/api/entities"
-	"github.com/ausocean/openfish/api/services"
+	"github.com/ausocean/openfish/cmd/openfish/api"
+	"github.com/ausocean/openfish/cmd/openfish/entities"
+	"github.com/ausocean/openfish/cmd/openfish/services"
 	"github.com/gofiber/fiber/v2"
 	"google.golang.org/api/idtoken"
 )
